move dependencies to integration package
