+++ token/sdk/sdk.go
@@ -38,7 +38,6 @@ import (
 	_ "github.com/hyperledger-labs/fabric-token-sdk/token/services/ttxdb/db/badger"
 	_ "github.com/hyperledger-labs/fabric-token-sdk/token/services/ttxdb/db/sql"
 	"github.com/pkg/errors"
-	_ "modernc.org/sqlite"
 )
 
 var logger = flogging.MustGetLogger("token-sdk")
